Added /map/reverse endpoint

+++ backend/CityShare.Backend.Application/Core/Models/Nominatim/Reverse/ReverseResultModel.cs
@@ -3,10 +3,10 @@
 
 public class ReverseResultModel
 {
-    public int place_id { get; set; }
+    public long place_id { get; set; }
     public string licence { get; set; } = default!;
     public string osm_type { get; set; } = default!;
-    public int osm_id { get; set; }
+    public long osm_id { get; set; }
     public string lat { get; set; } = default!;
     public string lon { get; set; } = default!;
     public string display_name { get; set; } = default!;
